Adição do primeiro teste, que já foi realizado, e correções decorrentes deste teste
O primeiro arquivo de testes do controlador foi feito, consistindo de
um sensor, que é um detector falso cujo circuito é um botão e um pull-down,
e um led como saída, além do timer DS1307. Assim, a adição de outros
processos de controle não foi testada ainda.

Durante a escrita e realização deste teste, diversos problemas e
inconveniências foram encontrados e corrigidos a medida que eles apareciam.
Assim, a classe BrewController e suas classes complementares estão
funcionais para o que é testado.

Ainda há problemas que não foram encontrados. O método moveMemTail precisa
de uma revisão e simplificação pois faz um processo de formatação que não
lhe cabe e, o principal problema encontrado, há um uso excessivo de memória
tanto por parte da BrewController quanto as outras classes, especialmente o
timer.

Esse problema prejudicou o cálculo do temporizador para ver se sua
contagem terminou. Porém, ele pôde ser corrigido por uma artimanha e o teste
pode ser concluído. Todavia, o problema em si permanece e impossibilita o
projeto final pois já pode-se ver que não haverá espaço para a parte de
interface com o usuário.

+++ brew-controller/src/brew-controller-lib/timer.h
@@ -17,9 +17,9 @@ protected:
 	boolean _time_over; //verdadeiro se a contagem chegou ao seu fim
 	boolean _active; //verdadeiro se o cronômetro está ativo, podendo ter sua contagem pausada
 	boolean _stopped; //verdadeiro se a contagem foi parada
-	long unsigned int _ref_time; //referência para cálculos de intervalos de tempo, em segundos
+	unsigned int _ref_time; //referência para cálculos de intervalos de tempo, em segundos
 	unsigned int _time_set; //intervalo de contagem do cronômetro
-	long unsigned int _time_when_stopped; //tempo em que o cronômetro parou, para que seja continuado
+	unsigned int _time_when_stopped; //tempo em que o cronômetro parou, para que seja continuado
 };
 
 #endif
